+++ ElnCoreModel/ELNModel/tblRefReactants.cs
@@ -9,7 +9,7 @@ public partial class tblRefReactants
 
     public string ProtocolItemID { get; set; } = null!;
 
-    public int SpecifiedUnitType { get; set; }
+    public short SpecifiedUnitType { get; set; }
 
     public byte IsDisplayAsVolume { get; set; }
